Consolidate process bootstrap for entrypoints
