+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "crates/braid-cli",
     "crates/braid-elaborate-js",
     "crates/braid-project",
+    "crates/braid-runtime",
     "crates/braid-governance",
     "crates/braid-vocab-rust",
     "crates/braid-manifest",
@@ -54,6 +55,7 @@ braid-verify = { path = "crates/braid-verify" }
 braid-render = { path = "crates/braid-render" }
 braid-sdk = { path = "crates/braid-sdk" }
 braid-manifest = { path = "crates/braid-manifest" }
+braid-runtime = { path = "crates/braid-runtime" }
 
 [profile.release]
 opt-level = 3
