+++ crates/braid-elaborate-js/tests/elaborate.rs
@@ -330,3 +330,20 @@ fn refusal_deep_nesting_fails_closed() {
     }
     assert!(matches!(elaborate_js(&deep), Err(ElabError::Parse(_))));
 }
+
+#[cfg(unix)]
+#[test]
+fn binary_rejects_non_utf8_startup_argument() {
+    use std::ffi::OsStr;
+    use std::os::unix::ffi::OsStrExt;
+    use std::process::Command;
+
+    let output = Command::new(env!("CARGO_BIN_EXE_braid-elaborate-js"))
+        .arg(OsStr::from_bytes(&[0xff]))
+        .output()
+        .expect("elaborator binary runs");
+    assert_eq!(output.status.code(), Some(2));
+    let stderr = String::from_utf8_lossy(&output.stderr);
+    assert!(stderr.contains("startup error"));
+    assert!(stderr.contains("not valid UTF-8"), "got: {stderr}");
+}
